requirements.txt
llama-index-embeddings-langchain==0.2.1
llama-index-llms-groq==0.2.0 
llama-index==0.11.20
llama-index-embeddings-huggingface==0.3.1
langchain-huggingface==0.1.0
gradio==5.4.0
python-dotenv